+++ Board-Main/src/main/java/org/jaeu/mapper/BoardMapper.java
@@ -7,7 +7,7 @@ import org.jaeu.domain.Criteria;
 
 public interface BoardMapper {
 	
-	public BoardVO detail(int bno);
+	public BoardVO detail(Long bno);
 	public List<BoardVO> getWithPaging(Criteria cri);
 	public void insertSelectKey(BoardVO board);
 	public int getTotal();
